refactor: payload validation code optimized
