+++ frontend/package.json
@@ -96,8 +96,8 @@
   },
   "lint-staged": {
      "**/*.{js,ts}": [
-      "npm run lint -- --fix",
-      "npm run format"
+      "npm run fix:eslint",
+      "prettier --write"
     ]
   }
 }
